+++ pakyow-core/lib/pakyow/behavior/commands.rb
@@ -10,7 +10,27 @@ module Pakyow
       extend Support::Extension
 
       apply_extension do
-        definable :command, Command
+        configurable :commands do
+          setting :paths, ["./commands", File.expand_path("../../commands", __FILE__)]
+          setting :prelaunch, []
+        end
+
+        definable :command, Command, builder: -> (*namespace, object_name, **opts) {
+          opts[:cli_name] = (namespace + [object_name]).join(":")
+
+          return namespace, object_name, opts
+        }, lookup: -> (_app, command, **values) {
+          command.call(**values)
+        }
+
+        # @api private
+        def load_commands
+          config.commands.paths.uniq.each_with_object(commands) do |commands_path, commands|
+            Dir.glob(File.join(File.expand_path(commands_path), "**/*.rb")).each do |command_path|
+              Kernel.load command_path
+            end
+          end
+        end
       end
     end
   end