+++ pakyow-core/lib/pakyow/cli/parsers/command.rb
@@ -28,20 +28,20 @@ module Pakyow
           unparsed = []
 
           OptionParser.new { |opts|
-            @command.flags.keys.each do |flag|
-              opts.on("--#{flag}") do |v|
-                options[flag] = v
+            @command.flags.each_key do |key|
+              opts.on("--#{key}") do |v|
+                options[key] = v
               end
             end
 
-            @command.options.keys.each do |option|
-              match = ["--#{option}=VAL"]
-              if @command.short_names.key?(option)
-                match.unshift("-#{@command.short_names[option]}VAL")
+            @command.options.each_pair do |key, option|
+              match = ["--#{key}=VAL"]
+              if short = option[:short]
+                match.unshift("-#{short}VAL")
               end
 
-              opts.on(*match) do |v|
-                options[option] = v
+              opts.on(*match) do |value|
+                options[key] = value
               end
             end
           }.order!(argv) do |arg|
@@ -49,21 +49,47 @@ module Pakyow
           end
 
           argv.concat(unparsed)
+
+          # Iterate over options again now that parsing has occurred.
+          #
+          @command.options.each_pair do |key, option|
+            resolve_default_value!(options, option, key)
+
+            if option[:required] && options[key].nil?
+              raise CLI::InvalidInput, "`#{key}' is a required option"
+            end
+          end
         rescue OptionParser::InvalidOption => error
           raise CLI::InvalidInput, "`#{error.args.first}' is not a supported option"
         end
 
-        def parse_arguments!(argv, options)
-          @command.arguments.each do |key, argument|
+        private def parse_arguments!(argv, options)
+          @command.arguments.each_pair do |key, argument|
             if argv.any?
               options[key] = argv.shift
-            elsif argument[:required]
+            end
+
+            resolve_default_value!(options, argument, key)
+
+            if argument[:required] && options[key].nil?
               raise CLI::InvalidInput, "`#{key}' is a required argument"
             end
           end
 
           if argv.any?
             raise CLI::InvalidInput, "`#{argv.shift}' is not a supported argument"
+          end
+        end
+
+        private def resolve_default_value!(options, option, key)
+          if !options.include?(key) && option.include?(:default)
+            value = option[:default]
+            options[key] = case value
+            when Proc
+              value.call
+            else
+              value
+            end
           end
         end
       end