+++ pakyow-core/spec/features/cli_spec.rb
@@ -62,7 +62,7 @@ RSpec.describe "command line interface" do
       expect(output).to include("\e[1mOPTIONS\e[0m\n")
       expect(output).to include("  -b, --baz=baz  \e[33mBaz arg\e[0m\e[31m (required)\e[0m\n")
       expect(output).to include("  -e, --env=env  \e[33mWhat environment to use\e[0m\n")
-      expect(output).to include("  -q, --qux=qux  \e[33mQux arg\e[0m\n")
+      expect(output).to include("  -q, --qux=qux  \e[33mQux arg (default: qux)\e[0m\n")
     end
 
     it "prints flags" do
@@ -215,15 +215,11 @@ RSpec.describe "command line interface" do
 
     # Run the command, capturing output.
     #
-    @output = capture_stdout do
-      allow($stdout).to receive(:tty?).and_return(true)
-      original_argv = ARGV.dup
-      ARGV.clear
-      ARGV.concat([command].concat(argv).compact)
-      eval(File.read(File.expand_path("../../../commands/pakyow", __FILE__)))
-      ARGV.clear
-      ARGV.concat(original_argv)
-    end
+    output = StringIO.new
+    allow(output).to receive(:tty?).and_return(true)
+    Pakyow::CLI.new([command].concat(argv).compact, feedback: Pakyow::CLI::Feedback.new(output))
+    output.rewind
+    @output = output.read
 
     # Set the working directory back to the original value.
     #