+++ pakyow-core/spec/unit/app/config_spec.rb
@@ -96,5 +96,11 @@ RSpec.describe Pakyow::Application do
         expect(app.config.tasks.prelaunch).to eq([])
       end
     end
+
+    describe "commands.prelaunch" do
+      it "exists" do
+        expect(app.config.commands.prelaunch).to eq([])
+      end
+    end
   end
 end