+++ pakyow-core/spec/unit/task_spec.rb
@@ -384,24 +384,24 @@ RSpec.describe Pakyow::Task do
     end
 
     it "returns help for the command" do
-      expect(instance.help).to eq("\e[34;1mTests help text\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow test_task\n")
+      expect(instance.help).to eq("\e[34;1mTests help text\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow test_task\n\n\e[1mOPTIONS\e[0m\n  -e, --env=env  \e[33mWhat environment to use\e[0m\n")
     end
 
     context "without the description" do
       it "returns help for the command" do
-        expect(instance.help(describe: false)).to eq("\n\e[1mUSAGE\e[0m\n  $ pakyow test_task\n")
+        expect(instance.help(describe: false)).to eq("\n\e[1mUSAGE\e[0m\n  $ pakyow test_task\n\n\e[1mOPTIONS\e[0m\n  -e, --env=env  \e[33mWhat environment to use\e[0m\n")
       end
     end
 
     context "with arguments" do
       let :arguments do
         {
-          qux: {
-            description: "qux arg"
-          },
           foo: {
             description: "foo arg",
             required: true
+          },
+          qux: {
+            description: "qux arg"
           }
         }
       end
@@ -411,7 +411,7 @@ RSpec.describe Pakyow::Task do
       end
 
       it "returns help for the command" do
-        expect(instance.help).to eq("\e[34;1mTests help text\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow test_task [FOO]\n\n\e[1mARGUMENTS\e[0m\n  FOO  \e[33mfoo arg\e[0m\e[31m (required)\e[0m\n  QUX  \e[33mqux arg\e[0m\n")
+        expect(instance.help).to eq("\e[34;1mTests help text\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow test_task [FOO]\n\n\e[1mARGUMENTS\e[0m\n  FOO  \e[33mfoo arg\e[0m\e[31m (required)\e[0m\n  QUX  \e[33mqux arg\e[0m\n\n\e[1mOPTIONS\e[0m\n  -e, --env=env  \e[33mWhat environment to use\e[0m\n")
       end
     end
 
@@ -420,12 +420,12 @@ RSpec.describe Pakyow::Task do
         {
           bar: {
             description: "bar arg",
-            short: :default
+            short: "b"
           },
           baz: {
             description: "baz arg",
             required: true,
-            short: :default
+            short: nil
           }
         }
       end
@@ -435,7 +435,7 @@ RSpec.describe Pakyow::Task do
       end
 
       it "returns help for the command" do
-        expect(instance.help).to eq("\e[34;1mTests help text\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow test_task --baz=baz\n\n\e[1mOPTIONS\e[0m\n  -b, --bar=bar  \e[33mbar arg\e[0m\n      --baz=baz  \e[33mbaz arg\e[0m\e[31m (required)\e[0m\n")
+        expect(instance.help).to eq("\e[34;1mTests help text\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow test_task --baz=baz\n\n\e[1mOPTIONS\e[0m\n  -e, --env=env  \e[33mWhat environment to use\e[0m\n  -b, --bar=bar  \e[33mbar arg\e[0m\n      --baz=baz  \e[33mbaz arg\e[0m\e[31m (required)\e[0m\n")
       end
     end
 
@@ -456,12 +456,12 @@ RSpec.describe Pakyow::Task do
         {
           bar: {
             description: "bar arg",
-            short: :default
+            short: "b"
           },
           baz: {
             description: "baz arg",
             required: true,
-            short: :default
+            short: nil
           }
         }
       end
@@ -471,7 +471,7 @@ RSpec.describe Pakyow::Task do
       end
 
       it "returns help for the command" do
-        expect(instance.help).to eq("\e[34;1mTests help text\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow test_task [FOO] --baz=baz\n\n\e[1mARGUMENTS\e[0m\n  FOO  \e[33mfoo arg\e[0m\e[31m (required)\e[0m\n  QUX  \e[33mqux arg\e[0m\n\n\e[1mOPTIONS\e[0m\n  -b, --bar=bar  \e[33mbar arg\e[0m\n      --baz=baz  \e[33mbaz arg\e[0m\e[31m (required)\e[0m\n")
+        expect(instance.help).to eq("\e[34;1mTests help text\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow test_task [FOO] --baz=baz\n\n\e[1mARGUMENTS\e[0m\n  FOO  \e[33mfoo arg\e[0m\e[31m (required)\e[0m\n  QUX  \e[33mqux arg\e[0m\n\n\e[1mOPTIONS\e[0m\n  -e, --env=env  \e[33mWhat environment to use\e[0m\n  -b, --bar=bar  \e[33mbar arg\e[0m\n      --baz=baz  \e[33mbaz arg\e[0m\e[31m (required)\e[0m\n")
       end
     end
   end